fix(openomni): preserve read back transport failures
