+++ packages/openomni/test/evidence/read-back-executor.test.ts
@@ -1,115 +1,18 @@
 /// <reference types="bun" />
 
 import { createHash } from "node:crypto";
-import { createServer, type Server, type ServerResponse } from "node:http";
 import { afterEach, describe, expect, test } from "bun:test";
-import { Storage, WorkItemStore } from "@openomni/session";
 import { ReadBackExecutor } from "../../src/index";
-
-const servers: Server[] = [];
-const hangingResponses: ServerResponse[] = [];
-const slowTimers: ReturnType<typeof setInterval>[] = [];
-const LOCAL_READ_BACK = { allowPrivateNetwork: true } satisfies ReadBackExecutor.Options;
-
-async function startFixtureServer(): Promise<string> {
-  const server = createServer((request, response) => {
-    const path = request.url ? new URL(request.url, "http://127.0.0.1").pathname : "/";
-    if (path === "/document") {
-      response.writeHead(200, { "content-type": "text/plain" });
-      response.end("The source contains a quoted passage for citation checks.");
-      return;
-    }
-    if (path === "/api/status") {
-      response.writeHead(200, { "content-type": "application/json" });
-      response.end(JSON.stringify({ ok: true, method: request.method }));
-      return;
-    }
-    if (path === "/binary") {
-      response.writeHead(200, { "content-type": "application/octet-stream" });
-      response.end(Buffer.from([0xff, 0xfe, 0xfd, 0x00, 0x61]));
-      return;
-    }
-    if (path === "/slow") {
-      response.writeHead(200, { "content-type": "text/plain" });
-      response.write("partial");
-      hangingResponses.push(response);
-      slowTimers.push(setInterval(() => response.write("x"), 5));
-      return;
-    }
-    if (path === "/large") {
-      response.writeHead(200, { "content-type": "text/plain" });
-      response.end("x".repeat(1_000_001));
-      return;
-    }
-    response.writeHead(404, { "content-type": "text/plain" });
-    response.end("missing");
-  });
-  await new Promise<void>((resolve, reject) => {
-    server.once("error", reject);
-    server.listen(0, "127.0.0.1", () => {
-      server.off("error", reject);
-      resolve();
-    });
-  });
-  const address = server.address();
-  if (!address || typeof address === "string") {
-    server.close();
-    throw new Error("fixture server did not bind to a TCP port");
-  }
-  servers.push(server);
-  return `http://127.0.0.1:${address.port}`;
-}
-
-function closeServer(server: Server): Promise<void> {
-  return new Promise((resolve, reject) => {
-    try {
-      server.close((error) => {
-        if (error) {
-          reject(error);
-          return;
-        }
-        resolve();
-      });
-    } catch (error) {
-      if (error instanceof Error && "code" in error && error.code === "ERR_SERVER_NOT_RUNNING") {
-        resolve();
-        return;
-      }
-      reject(error);
-    }
-  });
-}
-
-async function expectRejects(operation: Promise<unknown>): Promise<void> {
-  try {
-    await operation;
-  } catch (error) {
-    expect(error).toBeInstanceOf(Error);
-    return;
-  }
-  throw new Error("expected operation to reject");
-}
-
-async function createWorkItem() {
-  return WorkItemStore.create({
-    name: "read-back executor test",
-    sourceMessageId: "msg-readback",
-    sourceChannel: "test",
-    intent: "verify",
-    goal: "persist runtime read-back evidence",
-    acceptanceCriteria: ["The executor records external read-back evidence."],
-  });
-}
+import {
+  cleanupReadBackFixtures,
+  closeFixtureServers,
+  expectRejects,
+  LOCAL_READ_BACK,
+  startFixtureServer,
+} from "./read-back-fixture";
 
 afterEach(async () => {
-  Storage.reset();
-  for (const timer of slowTimers.splice(0)) {
-    clearInterval(timer);
-  }
-  for (const response of hangingResponses.splice(0)) {
-    response.destroy();
-  }
-  await Promise.all(servers.splice(0).map(closeServer));
+  await cleanupReadBackFixtures();
 });
 
 describe("ReadBackExecutor", () => {
@@ -157,7 +60,7 @@ describe("ReadBackExecutor", () => {
 
   test("marks a URL fetch as failed without digest when no response is observed", async () => {
     const origin = await startFixtureServer();
-    await Promise.all(servers.splice(0).map(closeServer));
+    await closeFixtureServers();
 
     const check = await ReadBackExecutor.execute(
       {
@@ -176,6 +79,35 @@ describe("ReadBackExecutor", () => {
     if (check.kind !== "url_fetch") throw new Error("expected url_fetch check");
     expect(check.statusCode).toBeUndefined();
     expect(check.contentDigest).toBeUndefined();
+  });
+
+  test("marks non-error transport rejections as failed read-back results", async () => {
+    const originalFetch = globalThis.fetch;
+    const failingFetch: typeof fetch = Object.assign(() => Promise.reject("transport-string"), {
+      preconnect: originalFetch.preconnect,
+    });
+    globalThis.fetch = failingFetch;
+
+    try {
+      const check = await ReadBackExecutor.execute(
+        {
+          kind: "url_fetch",
+          target: "http://127.0.0.1/document",
+        },
+        LOCAL_READ_BACK,
+      );
+
+      expect(check).toMatchObject({
+        kind: "url_fetch",
+        target: "http://127.0.0.1/document",
+        passed: false,
+      });
+      if (check.kind !== "url_fetch") throw new Error("expected url_fetch check");
+      expect(check.statusCode).toBeUndefined();
+      expect(check.contentDigest).toBeUndefined();
+    } finally {
+      globalThis.fetch = originalFetch;
+    }
   });
 
   test("uses timeoutMs as a wall-clock deadline", async () => {
@@ -334,35 +266,5 @@ describe("ReadBackExecutor", () => {
       statusCode: 200,
     });
     expect(check.matchedText).toBeUndefined();
-  });
-
-  test("persists runtime read-back evidence on a work item", async () => {
-    Storage.initialize({ dbPath: ":memory:" });
-    const origin = await startFixtureServer();
-    const item = await createWorkItem();
-
-    const updated = await ReadBackExecutor.record(
-      item.hash,
-      {
-        kind: "url_fetch",
-        target: `${origin}/document`,
-      },
-      LOCAL_READ_BACK,
-    );
-
-    const evidence = updated?.evidence.at(-1);
-    expect(evidence).toMatchObject({
-      kind: "verification",
-      passed: true,
-      readBack: {
-        kind: "url_fetch",
-        target: `${origin}/document`,
-        passed: true,
-        statusCode: 200,
-      },
-    });
-    const readBack = evidence?.readBack;
-    if (readBack?.kind !== "url_fetch") throw new Error("expected url_fetch evidence");
-    expect(readBack.contentDigest).toMatch(/^sha256:[a-f0-9]{64}$/);
   });
 });
